doc(Vite): make /dist deletable
Add documentation page for Vite example
