+++ examples/Vite/.gitignore
@@ -1,1 +1,3 @@
 cypress/support
+cypress/videos
+cypress/screenshots
