+++ examples/Vite/README.md
@@ -1,7 +1,7 @@
 itk-vite-example
 ===================
 
-This example demonstrates how to configure Vitk/Rollup to build a project that
+This example demonstrates how to configure Vite/Rollup to build a project that
 uses [itk-wasm](https://wasm.itk.org/).
 More information can be found in the [example
 documentation](https://wasm.itk.org/examples/vite.html).
@@ -15,10 +15,3 @@ npm run start
 
 And visit [http://localhost:8080/](http://localhost:8080/).
 
-## Development
-
-```
-npm install
-npm run build
-npm test
-```
