+++ examples/Vite/index.html
@@ -1,0 +1,21 @@
+<!DOCTYPE html>
+<html>
+  <head>
+    <title>itk-wasm Vite Example</title>
+    <meta http-equiv="Content-type" content="text/html; charset=utf-8" />
+  </head>
+
+  <body>
+    <!-- Input selector -->
+    <div>
+      <label>Select image or mesh:</label>
+      <input name="inputFile" type="file" />
+    </div>
+
+    <!-- File information -->
+    <textarea readonly name="fileInformation">File information...</textarea>
+
+    <!-- Javascript -->
+    <script type="module" src="./src/main.js"></script>
+  </body>
+</html>
