+++ examples/Vite/src/itkConfig.js
@@ -1,8 +1,8 @@
 const itkConfig = {
-  webWorkersUrl: '/itk/web-workers',
+  pipelineWorkerUrl: '/itk/web-workers/min-bundles/pipeline.worker.js',
   imageIOUrl: '/itk/image-io',
   meshIOUrl: '/itk/mesh-io',
-  pipelinesUrl: '/itk/pipelines',
+  pipelinesUrl: '/itk/pipelines'
 }
 
 export default itkConfig
