+++ examples/Vite/src/main.js
@@ -1,0 +1,6 @@
+import { outputFileInformation } from './outputFileInformation'
+
+const outputTextArea = document.querySelector('textarea')
+const handleFile = outputFileInformation(outputTextArea)
+const fileInput = document.querySelector('input')
+fileInput.addEventListener('change', handleFile)
